backend/authentication/pyproject.toml
[project]
name = "authentication"
version = "0.1.0"
description = "Authentication service"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt<4.0.0",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "redis[hiredis]>=7.0.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
